requirements.txt
google-generativeai>=0.8.0
elevenlabs>=1.0.0
mcp>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
